Refactoring stroies
